scripts/python/auto_start_proxy/requirements.txt
colorama==0.4.6
loguru==0.7.2
PyYAML==6.0.1
win32-setctime==1.1.0
